requirements.txt
opencv-python
numpy
pyyaml
torch
torchvision
requests
gradio
shap
grad-cam